// src/app/core/models/cliente.model.ts
export interface Cliente {
  id: string;
  cliente_numerodocumento: string;
  cliente_nombre: string;
  cliente_tipodocumento: string; //1: RUC
  cliente_direccion: string;
  cliente_pais: string;
  cliente_ciudad: string;
  cliente_codigoubigeo: string;
  cliente_departamento: string;
  cliente_provincia: string;
  cliente_distrito: string;
}